feat: smc 系统设置功能优化
