+++ core/src/main/java/com/tlcsdm/core/javafx/dialog/SystemSettingDialog.java
@@ -1,10 +1,9 @@
-package com.tlcsdm.smc.dialog;
+package com.tlcsdm.core.javafx.dialog;
 
 import com.tlcsdm.core.javafx.FxApp;
-import com.tlcsdm.core.javafx.dialog.FxDialog;
+import com.tlcsdm.core.javafx.controller.SystemSettingController;
 import com.tlcsdm.core.javafx.util.Config;
-import com.tlcsdm.smc.controller.SystemSettingController;
-import com.tlcsdm.smc.util.I18nUtils;
+import com.tlcsdm.core.javafx.util.I18nUtils;
 import javafx.scene.control.ButtonType;
 
 import java.util.ResourceBundle;
@@ -19,7 +18,7 @@ public class SystemSettingDialog {
         FxDialog<SystemSettingController> dialog = new FxDialog<SystemSettingController>()
                 .setResourceBundle(ResourceBundle.getBundle(I18nUtils.BASENAME, Config.defaultLocale))
                 .setTitle(title)
-                .setBodyFxml(SystemSettingDialog.class.getResource("/com/tlcsdm/smc/fxml/SystemSetting.fxml"))
+                .setBodyFxml(SystemSettingDialog.class.getResource("/com/tlcsdm/core/fxml/SystemSetting.fxml"))
                 .setOwner(FxApp.primaryStage)
                 .setButtonTypes(ButtonType.OK, ButtonType.CANCEL);
