+++ core/src/main/java/module-info.java
@@ -10,6 +10,9 @@ module com.tlcsdm.core {
     requires hutool.log;
     requires commons.beanutils;
 
+    opens com.tlcsdm.core.javafx.controller to javafx.fxml;
+    opens com.tlcsdm.core.javafx.view to javafx.fxml;
+
     exports com.tlcsdm.core.javafx;
     exports com.tlcsdm.core.javafx.control;
     exports com.tlcsdm.core.javafx.dialog;
