+++ core/src/main/resources/com/tlcsdm/core/i18n/messages_en.properties
@@ -5,3 +5,5 @@ core.buttonType.cancel=Cancel
 core.dialog.exception.dlg.title=Exception Details
 core.dialog.exception.dlg.header=Exception Details
 core.dialog.exception.dlg.label=The exception stacktrace was:
+core.dialog.systemSetting.check.exitShowAlert=Whether to show a confirmation dialog when exiting
+core.dialog.systemSetting.check.saveStageBound=Whether to save the window size and position information when exiting
