+++ core/src/main/resources/com/tlcsdm/core/i18n/messages_ja.properties
@@ -5,3 +5,5 @@ core.buttonType.cancel=\u30AD\u30E3\u30F3\u30BB\u30EB
 core.dialog.exception.dlg.title=\u4F8B\u5916\u306E\u8A73\u7D30
 core.dialog.exception.dlg.header=\u4F8B\u5916\u306E\u8A73\u7D30
 core.dialog.exception.dlg.label=\u4F8B\u5916\u306E\u30B9\u30BF\u30C3\u30AF \u30C8\u30EC\u30FC\u30B9\u306F\u6B21\u306E\u3068\u304A\u308A\u3067\u3059\u3002
+core.dialog.systemSetting.check.exitShowAlert=\u7D42\u4E86\u6642\u306B\u78BA\u8A8D\u30C0\u30A4\u30A2\u30ED\u30B0\u3092\u8868\u793A\u3059\u308B\u304B\u3069\u3046\u304B
+core.dialog.systemSetting.check.saveStageBound=\u7D42\u4E86\u6642\u306B\u30A6\u30A3\u30F3\u30C9\u30A6\u306E\u30B5\u30A4\u30BA\u3068\u4F4D\u7F6E\u60C5\u5831\u3092\u4FDD\u5B58\u3059\u308B\u304B\u3069\u3046\u304B
