+++ core/src/main/resources/com/tlcsdm/core/i18n/messages_zh.properties
@@ -5,3 +5,5 @@ core.buttonType.cancel=\u53D6\u6D88
 core.dialog.exception.dlg.title=\u5F02\u5E38\u8BE6\u60C5
 core.dialog.exception.dlg.header=\u5F02\u5E38\u8BE6\u60C5
 core.dialog.exception.dlg.label=\u5F02\u5E38\u5806\u6808\u4FE1\u606F:
+core.dialog.systemSetting.check.exitShowAlert=\u9000\u51FA\u65F6\u662F\u5426\u663E\u793A\u786E\u8BA4\u5BF9\u8BDD\u6846
+core.dialog.systemSetting.check.saveStageBound=\u662F\u5426\u5728\u9000\u51FA\u65F6\u4FDD\u5B58\u7A97\u53E3\u5927\u5C0F\u53CA\u4F4D\u7F6E\u4FE1\u606F
